+++ React/tic-tac/src/config/api.js
@@ -1,0 +1,8 @@
+import { create } from 'apisauce';
+
+const api = create({
+  baseURL: process.env.REACT_APP_API_BASE_URL,
+  timeout: 5000
+});
+
+export default api;
