+++ React/tic-tac/src/redux/auth/actions.js
@@ -1,28 +1,7 @@
-import UserService from '../../services/UserService';
-
-export function LogInTrue(payload) {
+function LogInTrue(payload) {
   return {
     type: 'LogIn',
     loggedIn: payload
-  };
-}
-
-export function LogIn(payload) {
-  return dispatch => {
-    UserService.post('/Users/login', {
-      email: payload.email,
-      password: payload.password
-    })
-      .then(response => {
-        if (response.ok) {
-          dispatch(LogInTrue(true));
-        } else {
-          throw new Error();
-        }
-      })
-      .catch(() => {
-        alert('Error al iniciar sesión');
-      });
   };
 }
 
@@ -31,3 +10,5 @@ export function LogOut() {
     type: 'LogOut'
   };
 }
+
+export default LogInTrue;
