--- React/tic-tac/src/redux/form/actions.js
@@ -1,1 +1,0 @@
-import getLogin from '../../services/DataServices';
